Using approriate data type for loop variable

+++ main.cpp
@@ -258,7 +258,7 @@ int main(int argc, char *argv[])
                 static const size_t combined_len = 8188;
                 char combined[combined_len + 4];
                 for (int s = 3; s >= 1; --s) {
-                    for (unsigned int i = 0; i <= words.size() - s; ++i) {
+                    for (size_t i = 0; i <= words.size() - s; ++i) {
                         char *p = combined;
                         for (int k = 0; k < s; ++k) {
                             if (k > 0)
